made some changes and working good0
