+++ src/components/views/StudentView.jsx
@@ -1,0 +1,25 @@
+import React from 'react';
+
+const StudentsView = ({ renderFilters, renderStudentList, onAddStudent, onFilterChange }) => (
+    <div className="space-y-6">
+      <div className="flex justify-between items-center mb-8">
+        <div>
+          <h1 className="text-2xl font-bold text-gray-900">Students</h1>
+          <p className="mt-1 text-sm text-gray-500">Manage your students and their information</p>
+        </div>
+        <button
+          onClick={onAddStudent}
+          className="inline-flex items-center px-4 py-2 border border-transparent shadow-sm text-sm font-medium rounded-lg text-white bg-purple-600 hover:bg-purple-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-purple-500 transition-colors"
+        >
+          <svg className="-ml-1 mr-2 h-5 w-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 6v6m0 0v6m0-6h6m-6 0H6" />
+          </svg>
+          Add New Student
+        </button>
+      </div>
+      {renderFilters(onFilterChange)}
+      {renderStudentList()}
+    </div>
+  );
+
+  export default StudentsView;
